pyproject.toml
[tool.poetry]
name = "discord-memo"
version = "0.1.0"
description = ""
authors = ["kamesan1577 <kamesan1577@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.2"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
